Add low level python packager
Port over the python dependency builder from Chalice. Also added a small
wrapper class to adapt the interface to something more like what the
Builder Action will use. The interface for the low level will stay the
same, the wrapper PythonPipDependencyBuilder class can be modified as
needed to match the expected interface as it evolves.