Release: v0.2.0 (#95)
* Fix local package installation (#72)

When packaging the dependencies only the first local package would be
included in the final bundle. This change adds support for any number
of local directory links to be treated as buildable dependencies. Tests
were also added for the case of 1 or 2 local directory links specified
in the requirements.txt.

* fix: Remove tests from .whl file (bdist_wheel) (#90)

* Add .NET Core builder support (#91)

* feat(maven): Support for Java Maven builder (#87)

* chore: version bump to 0.2.0 (#94)