fix: Remove MODULE_NAME command from Maven Workflow
On Windows, the echo command is not valid in most cases. The Maven Workflow
was doing some operations that help support the multiple project build
use-case. Given this sometimes fails on Windows, the PR updates the Maven
Worklfow to only install and copy dependencies. This removes the need
to echo and therefore allows Windows Maven builds to succeed without
needing to run within the container.