fix: WindowsError keyword is undefined on non-Windows platforms (#117)
WindowsError is a subclass of `OSError`. It is defined and raised only
in Windows. Hence the exception statement will return a
`WindowsError keyword is undefined` in other platforms.

Usually this error is encountered when CodeUri: /folder/does/not/exist. The builder tries to copy the files and fails because the path does not exist. On Windows, the exception handler works, but on Mac/Linux Python interpreter does not find the variable WindowsError and crashes. So it leads to a vague WindowsError on Mac/Linux.