chore: Move to ruff from pylint (#435)
When we started the project we defaulted to use pylint. Pylint has served it's
purpose but it pretty slow. Ruff is a newer linter in the python
ecosystem but is written in Rust. This makes Ruff was faster than pylint.
On my machine (while testing in SAM CLI), pylint took about 70s to lint the repo
but with ruff it took .04s.

Co-authored-by: Jacob Fuss <jfuss@users.noreply.github.com>